Removed redundant function 'newchain'
The removed function 'newchain(a, b)' was almost the same as calling
!chain_equal(a, b), varying slightly only in the amount of data compared
in the non-fixed-length data portion of given chain nodes.

A third argument 'data_size' was introduced into 'chain_equal' function
in order to allow it to know how many bytes to compare in the
variable-length data portion of the chain nodes.

A macro 'CHAIN_LENGTH(e)' was introduced to allow to easy calculate the
total length of the non-fixed-length data part of chain nodes.

Check the thread below for more details:
https://gitlab.isc.org/isc-projects/bind9/-/merge_requests/291#note_12184