store all netmgr timeouts is millisecond units
previously the TCP timeouts set by isc_nm_tcp_settimeouts() and
used in the keepalive option were stored in units of 100
milliseconds.  that is inconsistent with the values used in
isc_nm_*connect() functions and in the new isc_nm_settimeout()
function.  this commit changes all timeouts to use milliseconds
and updates callers of the old functions to normalize the values
as necessary.