Add test case for going from signed to insecure
When going insecure, first make sure that CDS and CDNSKEY DELETE
records are published. Add this to the check_ds function so that
for other policies we can validate that these records are not present
in the response.

The second step of going insecure is to wait until the DS becomes
HIDDEN. Then it is safe to remove the rest of the DNSSEC records from
the zone (DNSKEY, RRSIG).