Update keymgr to allow transition to insecure mode
The keymgr prevented zones from going to insecure mode. If we
have a policy with an empty key list this is an explicit signal that
the operator wants to go back to insecure mode. In this case allow
one extra state transition to be valid when checking for DNSSEC
safety.