Treat zone going insecure as inline-signing
When a zone removes the "dnssec-policy" from its zone configuration,
the keymgr should still run to allow for a graceful transition to
insecure. So if we do not (no longer) have a DNSSEC policy attached
to the zone, check for key state files. If there exists at least one
take this as an indication that the zone goes back to insecure.

Keep running the keymgr, so that the zone transitions to insecure
without going bogus.

This commit ensures that during reconfiguration the zone is kept
being treated as an inline-signing zone.